(BSR)[PRO] chore: add AppRouter component
