+++ pro/src/app/AppRouter/AppRouter.tsx
@@ -1,0 +1,85 @@
+import React, { useEffect, useState } from 'react'
+import { useSelector } from 'react-redux'
+import { Redirect, Route, Switch } from 'react-router'
+
+import AppLayout from 'app/AppLayout'
+import useActiveFeature from 'components/hooks/useActiveFeature'
+import Spinner from 'components/layout/Spinner'
+import NotFound from 'components/pages/Errors/NotFound/NotFound'
+import {
+  selectActiveFeatures,
+  selectFeaturesInitialized,
+} from 'store/features/selectors'
+import routes, { IRoute, routesWithoutLayout } from 'utils/routes_map'
+
+const AppRouter = (): JSX.Element => {
+  const isFeaturesInitialized = useSelector(selectFeaturesInitialized)
+  const activeFeatures = useSelector(selectActiveFeatures)
+  const [activeRoutes, setActiveRoutes] = useState<IRoute[]>([])
+  const [activeRoutesWithoutLayout, setActiveRoutesWithoutLayout] = useState<
+    IRoute[]
+  >([])
+  const useSummaryPage = useActiveFeature('OFFER_FORM_SUMMARY_PAGE')
+
+  useEffect(() => {
+    setActiveRoutes(
+      routes.filter(
+        route =>
+          !route.featureName || activeFeatures.includes(route.featureName)
+      )
+    )
+
+    setActiveRoutesWithoutLayout(
+      routesWithoutLayout.filter(
+        route =>
+          !route.featureName || activeFeatures.includes(route.featureName)
+      )
+    )
+  }, [activeFeatures])
+
+  if (!isFeaturesInitialized) {
+    return (
+      <main className="spinner-container">
+        <Spinner />
+      </main>
+    )
+  }
+
+  return (
+    <Switch>
+      <Redirect
+        from="/offres/:offerId([A-Z0-9]+)/edition"
+        to={
+          useSummaryPage
+            ? '/offre/:offerId([A-Z0-9]+)/individuel/recapitulatif'
+            : '/offre/:offerId([A-Z0-9]+)/individuel/edition'
+        }
+      />
+      <Redirect
+        from="/offre/:offerId([A-Z0-9]+)/scolaire/edition"
+        to="/offre/:offerId([A-Z0-9]+)/collectif/edition"
+      />
+      {activeRoutes.map(route => (
+        <Route
+          exact={route.exact}
+          key={Array.isArray(route.path) ? route.path.join('|') : route.path}
+          path={route.path}
+        >
+          <AppLayout layoutConfig={route.meta && route.meta.layoutConfig}>
+            <route.component />
+          </AppLayout>
+        </Route>
+      ))}
+      {activeRoutesWithoutLayout.map(route => (
+        <Route
+          {...route}
+          exact={route.exact}
+          key={Array.isArray(route.path) ? route.path.join('|') : route.path}
+        />
+      ))}
+      <Route component={NotFound} />
+    </Switch>
+  )
+}
+
+export default AppRouter
