+++ pro/src/app/AppRouter/index.ts
@@ -1,0 +1,1 @@
+export { default as AppRouter } from './AppRouter'
